Added controller for advanced search, added findIntersection() and compareTo() methods. Changed signature of caloriesBetween() method

+++ Nutty/src/main/java/cmpesweng2014/group1/nutty/dao/RecipeDao.java
@@ -455,7 +455,7 @@ public class RecipeDao extends PcDao{
 			return recipes;		
 		}
 	}
-	public Recipe[] caloriesBetween(int upperlimit, int lowerlimit){
+	public Recipe[] caloriesBetween(double upperlimit, double lowerlimit){
 		List<Recipe> recList = this.getTemplate().query(
 				"SELECT * FROM Recipe WHERE total_calorie > ? AND total_calorie < ? ",
 				new Object[] { lowerlimit, upperlimit }, new RecipeRowMapper());
